merge into skynode master branch

+++ ElectronNET.CLI/Commands/BuildCommand.cs
@@ -1,7 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using System.Runtime.InteropServices;
+using System.Text;
 using System.Threading.Tasks;
 using ElectronNET.CLI.Commands.Actions;
 
@@ -11,7 +13,14 @@ namespace ElectronNET.CLI.Commands
     {
         public const string COMMAND_NAME = "build";
         public const string COMMAND_DESCRIPTION = "Build your Electron Application.";
-        public const string COMMAND_ARGUMENTS = "<Platform> to build (default is current OS, possible values are: win,osx,linux)";
+        public static string COMMAND_ARGUMENTS = "Needed: '/target' with params 'win/osx/linux' to build for a typical app or use 'custom' and specify .NET Core build config & electron build config" + Environment.NewLine +
+                                                 " for custom target, check .NET Core RID Catalog and Electron build target/" + Environment.NewLine +
+                                                 " e.g. '/target win' or '/target custom \"win7-x86;win32\"'" + Environment.NewLine +
+                                                 "Optional: '/dotnet-configuration' with the desired .NET Core build config e.g. release or debug. Default = Release" + Environment.NewLine + 
+                                                 "Optional: '/electron-arch' to specify the resulting electron processor architecture (e.g. ia86 for x86 builds). Be aware to use the '/target custom' param as well!" + Environment.NewLine +
+                                                 "Optional: '/electron-params' specify any other valid parameter, which will be routed to the electron-packager." + Environment.NewLine +
+                                                 "Full example for a 32bit debug build with electron prune: build /target custom win7-x86;win32 /dotnet-configuration Debug /electron-arch ia32  /electron-params \"--prune=true \"";
+
         public static IList<CommandOption> CommandOptions { get; set; } = new List<CommandOption>();
 
         private string[] _args;
@@ -21,20 +30,36 @@ namespace ElectronNET.CLI.Commands
             _args = args;
         }
 
+        private string _paramTarget = "target";
+        private string _paramDotNetConfig = "dotnet-configuration";
+        private string _paramElectronArch = "electron-arch";
+        private string _paramElectronParams = "electron-params";
+
         public Task<bool> ExecuteAsync()
         {
             return Task.Run(() =>
             {
                 Console.WriteLine("Build Electron Application...");
 
-                string desiredPlatform = "";
+                SimpleCommandLineParser parser = new SimpleCommandLineParser();
+                parser.Parse(_args);
 
-                if (_args.Length > 0)
+                var desiredPlatform = parser.Arguments[_paramTarget][0];
+                string specifiedFromCustom = string.Empty;
+                if (desiredPlatform == "custom" && parser.Arguments[_paramTarget].Length > 1)
                 {
-                    desiredPlatform = _args[0];
+                    specifiedFromCustom = parser.Arguments[_paramTarget][1];
                 }
 
-                var platformInfo = GetTargetPlatformInformation.Do(desiredPlatform);
+                string configuration = "Release";
+                if (parser.Arguments.ContainsKey(_paramDotNetConfig))
+                {
+                    configuration = parser.Arguments[_paramDotNetConfig][0];
+                }
+
+                var platformInfo = GetTargetPlatformInformation.Do(desiredPlatform, specifiedFromCustom);
+
+                Console.WriteLine($"Build ASP.NET Core App for {platformInfo.NetCorePublishRid}...");
 
 
                 string tempPath = Path.Combine(Directory.GetCurrentDirectory(), "obj", "desktop", desiredPlatform);
@@ -47,27 +72,44 @@ namespace ElectronNET.CLI.Commands
 
                 string tempBinPath = Path.Combine(tempPath, "bin");
 
-                Console.WriteLine($"Build ASP.NET Core App for {platformInfo.NetCorePublishRid}...");
+                Console.WriteLine($"Build ASP.NET Core App for {platformInfo.NetCorePublishRid} under {configuration}-Configuration...");
 
-                ProcessHelper.CmdExecute($"dotnet publish -r {platformInfo.NetCorePublishRid} --output \"{tempBinPath}\"", Directory.GetCurrentDirectory());
+                var resultCode = ProcessHelper.CmdExecute($"dotnet publish -r {platformInfo.NetCorePublishRid} -c {configuration} --output \"{tempBinPath}\"", Directory.GetCurrentDirectory());
+
+                if (resultCode != 0)
+                {
+                    Console.WriteLine("Error occurred during dotnet publish.");
+                    return false;
+                }
 
                 DeployEmbeddedElectronFiles.Do(tempPath);
 
-                Console.WriteLine("Start npm install...");
-                ProcessHelper.CmdExecute("npm install", tempPath);
+                var checkForNodeModulesDirPath = Path.Combine(tempPath, "node_modules");
 
-                Console.WriteLine("Start npm install electron-packager...");
-
-                if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+                if (Directory.Exists(checkForNodeModulesDirPath) == false)
                 {
-                    // Works proper on Windows... 
-                    ProcessHelper.CmdExecute("npm install electron-packager --global", tempPath);
+                    Console.WriteLine("node_modules missing in: " + checkForNodeModulesDirPath);
+
+                    Console.WriteLine("Start npm install...");
+                    ProcessHelper.CmdExecute("npm install", tempPath);
+
+                    Console.WriteLine("Start npm install electron-packager...");
+
+                    if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+                    {
+                        // Works proper on Windows... 
+                        ProcessHelper.CmdExecute("npm install electron-packager --global", tempPath);
+                    }
+                    else
+                    {
+                        // ToDo: find another solution or document it proper
+                        // GH Issue https://github.com/electron-userland/electron-prebuilt/issues/48
+                        Console.WriteLine("Electron Packager - make sure you invoke 'sudo npm install electron-packager --global' at " + tempPath + " manually. Sry.");
+                    }
                 }
                 else
                 {
-                    // ToDo: find another solution or document it proper
-                    // GH Issue https://github.com/electron-userland/electron-prebuilt/issues/48
-                    Console.WriteLine("Electron Packager - make sure you invoke 'sudo npm install electron-packager --global' at " + tempPath + " manually. Sry.");
+                    Console.WriteLine("Skip npm install, because node_modules directory exists in: " + checkForNodeModulesDirPath);
                 }
 
                 Console.WriteLine("Build Electron Desktop Application...");
@@ -76,8 +118,21 @@ namespace ElectronNET.CLI.Commands
                 Console.WriteLine("Executing electron magic in this directory: " + buildPath);
 
                 // ToDo: Need a solution for --asar support
+
+                string electronArch = "x64";
+                if (parser.Arguments.ContainsKey(_paramElectronArch))
+                {
+                    electronArch = parser.Arguments[_paramElectronArch][0];
+                }
+
+                string electronParams = "";
+                if (parser.Arguments.ContainsKey(_paramElectronParams))
+                {
+                    electronParams = parser.Arguments[_paramElectronParams][0];
+                }
+
                 Console.WriteLine($"Package Electron App for Platform {platformInfo.ElectronPackerPlatform}...");
-                ProcessHelper.CmdExecute($"electron-packager . --platform={platformInfo.ElectronPackerPlatform} --arch=x64 --out=\"{buildPath}\" --overwrite", tempPath);
+                ProcessHelper.CmdExecute($"electron-packager . --platform={platformInfo.ElectronPackerPlatform} --arch={electronArch} {electronParams} --out=\"{buildPath}\" --overwrite", tempPath);
 
                 Console.WriteLine("... done");
